Un-unify tempo and time sequences...
Add an update_tempomap() method to Timeline instead, which must be
called every time the time or tempo tracks are edited. This is perhaps as
ugly as the unified map was, but is at least a bit more compatible with
the possibility of moving all the meat of tempomapping into a library.